Fix Visual Studio build by providing custom Exiv2 recipe
There is an issue with the official conan recipe for Exiv2. that attempt
to buid a GNU specific library on the windows platform.

It is actually possible to build this library on Windows, but require
emulating GNU via cygwin or some other means. The developers of Exiv2
recommends against doing so. This library is only used for international
conversion of character sets and is not a required dependancy of the
library.

This commit copies the existing exiv2 conan recipe locally, and add a
couple of `if self.settings.os != "Windows"` on the commands that
include the libiconv package.

This will fix #18 for now. But a proper fix would be to get these
changes upstream so they are included in the conan index.

I do not know at this time if the conanfile for Exiv2 is provided by the
Exiv2 project or a third party.

If I submit a patch to fix the conan package upstream, and it get
accepted, we may be able to undo the changes provided here.