Fix usage of ⌘ key in tooltips on non Apple platforms
This patch introduce a new header called osMacros.h.

This header is intended to contain build-time information about the
operating system detected by the compiler, and provide whatever may be
platform dependant.

For now it only has the switch for the name of the keyboard key you use
for some operations, which is Cmd (⌘) on Mac and Ctrl on all the other
computer in the known universe 😜